put between code in compare file Co-authored-by: Sutou Kouhei <kou@cozmixng.org>
