+++ cpp/src/arrow/CMakeLists.txt
@@ -422,7 +422,6 @@ if(ARROW_COMPUTE)
        compute/kernels/hash_aggregate.cc
        compute/kernels/row_encoder.cc
        compute/kernels/scalar_arithmetic.cc
-       compute/kernels/scalar_between.cc
        compute/kernels/scalar_boolean.cc
        compute/kernels/scalar_cast_boolean.cc
        compute/kernels/scalar_cast_dictionary.cc
