+++ cpp/src/arrow/compute/api_scalar.h
@@ -1653,4 +1653,3 @@ Result<Datum> Between(const Datum& values, const Datum& left, const Datum& right
 
 }  // namespace compute
 }  // namespace arrow
-
