+++ cpp/src/arrow/compute/kernels/scalar_compare.cc
@@ -34,6 +34,10 @@ using util::string_view;
 namespace compute {
 namespace internal {
 
+using applicator::ScalarBinary;
+using applicator::ScalarBinaryEqualTypes;
+using applicator::ScalarTernaryEqualTypes;
+
 namespace {
 
 struct Equal {
@@ -65,6 +69,17 @@ struct GreaterEqual {
   static constexpr T Call(KernelContext*, const Arg0& left, const Arg1& right, Status*) {
     static_assert(std::is_same<T, bool>::value && std::is_same<Arg0, Arg1>::value, "");
     return left >= right;
+  }
+};
+
+struct Between {
+  template <typename T, typename Arg0, typename Arg1, typename Arg2>
+  static constexpr T Call(KernelContext*, const Arg0& middle, const Arg1& left,
+                          const Arg2& right, Status*) {
+    static_assert(std::is_same<T, bool>::value && std::is_same<Arg0, Arg1>::value &&
+                      std::is_same<Arg1, Arg2>::value,
+                  "");
+    return (left <= middle) && (middle <= right);
   }
 };
 
@@ -159,9 +174,8 @@ struct Maximum {
 // Implement Less, LessEqual by flipping arguments to Greater, GreaterEqual
 
 template <typename OutType, typename ArgType, typename Op>
-struct CompareTimestamps
-    : public applicator::ScalarBinaryEqualTypes<OutType, ArgType, Op> {
-  using Base = applicator::ScalarBinaryEqualTypes<OutType, ArgType, Op>;
+struct CompareTimestamps : public ScalarBinaryEqualTypes<OutType, ArgType, Op> {
+  using Base = ScalarBinaryEqualTypes<OutType, ArgType, Op>;
 
   static Status Exec(KernelContext* ctx, const ExecSpan& batch, ExecResult* out) {
     const auto& lhs = checked_cast<const TimestampType&>(*batch[0].type());
@@ -177,17 +191,18 @@ struct CompareTimestamps
 
 template <typename Op>
 void AddIntegerCompare(const std::shared_ptr<DataType>& ty, ScalarFunction* func) {
-  auto exec =
-      GeneratePhysicalInteger<applicator::ScalarBinaryEqualTypes, BooleanType, Op>(*ty);
+  auto exec = GeneratePhysicalInteger<ScalarBinaryEqualTypes, BooleanType, Op>(*ty);
   DCHECK_OK(func->AddKernel({ty, ty}, boolean(), std::move(exec)));
 }
 
 template <typename InType, typename Op>
 void AddGenericCompare(const std::shared_ptr<DataType>& ty, ScalarFunction* func) {
-  DCHECK_OK(
-      func->AddKernel({ty, ty}, boolean(),
-                      applicator::ScalarBinaryEqualTypes<BooleanType, InType, Op>::Exec));
+  DCHECK_OK(func->AddKernel({ty, ty}, boolean(),
+                            ScalarBinaryEqualTypes<BooleanType, InType, Op>::Exec));
 }
+
+// Cast to appropriate type if needed and examine
+// for missing values
 
 struct CompareFunction : ScalarFunction {
   using ScalarFunction::ScalarFunction;
@@ -243,9 +258,9 @@ template <typename Op>
 std::shared_ptr<ScalarFunction> MakeCompareFunction(std::string name, FunctionDoc doc) {
   auto func = std::make_shared<CompareFunction>(name, Arity::Binary(), std::move(doc));
 
-  DCHECK_OK(func->AddKernel(
-      {boolean(), boolean()}, boolean(),
-      applicator::ScalarBinary<BooleanType, BooleanType, BooleanType, Op>::Exec));
+  DCHECK_OK(
+      func->AddKernel({boolean(), boolean()}, boolean(),
+                      ScalarBinary<BooleanType, BooleanType, BooleanType, Op>::Exec));
 
   for (const std::shared_ptr<DataType>& ty : IntTypes()) {
     AddIntegerCompare<Op>(ty, func.get());
@@ -266,43 +281,35 @@ std::shared_ptr<ScalarFunction> MakeCompareFunction(std::string name, FunctionDo
   // Duration
   for (auto unit : TimeUnit::values()) {
     InputType in_type(match::DurationTypeUnit(unit));
-    auto exec =
-        GeneratePhysicalInteger<applicator::ScalarBinaryEqualTypes, BooleanType, Op>(
-            int64());
+    auto exec = GeneratePhysicalInteger<ScalarBinaryEqualTypes, BooleanType, Op>(int64());
     DCHECK_OK(func->AddKernel({in_type, in_type}, boolean(), std::move(exec)));
   }
 
   // Time32 and Time64
   for (auto unit : {TimeUnit::SECOND, TimeUnit::MILLI}) {
     InputType in_type(match::Time32TypeUnit(unit));
-    auto exec =
-        GeneratePhysicalInteger<applicator::ScalarBinaryEqualTypes, BooleanType, Op>(
-            int32());
+    auto exec = GeneratePhysicalInteger<ScalarBinaryEqualTypes, BooleanType, Op>(int32());
     DCHECK_OK(func->AddKernel({in_type, in_type}, boolean(), std::move(exec)));
   }
   for (auto unit : {TimeUnit::MICRO, TimeUnit::NANO}) {
     InputType in_type(match::Time64TypeUnit(unit));
-    auto exec =
-        GeneratePhysicalInteger<applicator::ScalarBinaryEqualTypes, BooleanType, Op>(
-            int64());
+    auto exec = GeneratePhysicalInteger<ScalarBinaryEqualTypes, BooleanType, Op>(int64());
     DCHECK_OK(func->AddKernel({in_type, in_type}, boolean(), std::move(exec)));
   }
 
   for (const std::shared_ptr<DataType>& ty : BaseBinaryTypes()) {
-    auto exec =
-        GenerateVarBinaryBase<applicator::ScalarBinaryEqualTypes, BooleanType, Op>(*ty);
+    auto exec = GenerateVarBinaryBase<ScalarBinaryEqualTypes, BooleanType, Op>(*ty);
     DCHECK_OK(func->AddKernel({ty, ty}, boolean(), std::move(exec)));
   }
 
   for (const auto id : {Type::DECIMAL128, Type::DECIMAL256}) {
-    auto exec = GenerateDecimal<applicator::ScalarBinaryEqualTypes, BooleanType, Op>(id);
+    auto exec = GenerateDecimal<ScalarBinaryEqualTypes, BooleanType, Op>(id);
     DCHECK_OK(
         func->AddKernel({InputType(id), InputType(id)}, boolean(), std::move(exec)));
   }
 
   {
-    auto exec =
-        applicator::ScalarBinaryEqualTypes<BooleanType, FixedSizeBinaryType, Op>::Exec;
+    auto exec = ScalarBinaryEqualTypes<BooleanType, FixedSizeBinaryType, Op>::Exec;
     auto ty = InputType(Type::FIXED_SIZE_BINARY);
     DCHECK_OK(func->AddKernel({ty, ty}, boolean(), std::move(exec)));
   }
@@ -699,6 +706,68 @@ std::shared_ptr<ScalarFunction> MakeScalarMinMax(std::string name, FunctionDoc d
   return func;
 }
 
+template <typename Op>
+void AddIntegerBetween(const std::shared_ptr<DataType>& ty, ScalarFunction* func) {
+  auto exec = GeneratePhysicalInteger<ScalarTernaryEqualTypes, BooleanType, Op>(*ty);
+  DCHECK_OK(func->AddKernel({ty, ty, ty}, boolean(), std::move(exec)));
+}
+
+template <typename InType, typename Op>
+void AddGenericBetween(const std::shared_ptr<DataType>& ty, ScalarFunction* func) {
+  DCHECK_OK(func->AddKernel({ty, ty, ty}, boolean(),
+                            ScalarTernaryEqualTypes<BooleanType, InType, Op>::Exec));
+}
+
+template <typename Op>
+std::shared_ptr<ScalarFunction> MakeBetweenFunction(std::string name,
+                                                    const FunctionDoc* doc) {
+  auto func = std::make_shared<CompareFunction>(name, Arity::Ternary(), doc);
+
+  for (const std::shared_ptr<DataType>& ty : IntTypes()) {
+    AddIntegerBetween<Op>(ty, func.get());
+  }
+
+  AddIntegerBetween<Op>(date32(), func.get());
+  AddIntegerBetween<Op>(date64(), func.get());
+
+  AddGenericBetween<FloatType, Op>(float32(), func.get());
+  AddGenericBetween<DoubleType, Op>(float64(), func.get());
+
+  // Duration
+  for (auto unit : TimeUnit::values()) {
+    InputType in_type(match::DurationTypeUnit(unit));
+    auto exec =
+        GeneratePhysicalInteger<ScalarTernaryEqualTypes, BooleanType, Op>(int64());
+    DCHECK_OK(func->AddKernel({in_type, in_type, in_type}, boolean(), std::move(exec)));
+  }
+
+  // Time32 and Time64
+  for (auto unit : {TimeUnit::SECOND, TimeUnit::MILLI}) {
+    InputType in_type(match::Time32TypeUnit(unit));
+    auto exec =
+        GeneratePhysicalInteger<ScalarTernaryEqualTypes, BooleanType, Op>(int32());
+    DCHECK_OK(func->AddKernel({in_type, in_type, in_type}, boolean(), std::move(exec)));
+  }
+  for (auto unit : {TimeUnit::MICRO, TimeUnit::NANO}) {
+    InputType in_type(match::Time64TypeUnit(unit));
+    auto exec =
+        GeneratePhysicalInteger<ScalarTernaryEqualTypes, BooleanType, Op>(int64());
+    DCHECK_OK(func->AddKernel({in_type, in_type, in_type}, boolean(), std::move(exec)));
+  }
+
+  for (const std::shared_ptr<DataType>& ty : BaseBinaryTypes()) {
+    auto exec = GenerateVarBinaryBase<ScalarTernaryEqualTypes, BooleanType, Op>(*ty);
+    DCHECK_OK(func->AddKernel(
+        {InputType::Array(ty), InputType::Scalar(ty), InputType::Scalar(ty)}, boolean(),
+        exec));
+    DCHECK_OK(func->AddKernel(
+        {InputType::Array(ty), InputType::Array(ty), InputType::Array(ty)}, boolean(),
+        std::move(exec)));
+  }
+
+  return func;
+}
+
 const FunctionDoc equal_doc{"Compare values for equality (x == y)",
                             ("A null on either side emits a null comparison result."),
                             {"x", "y"}};
@@ -739,6 +808,10 @@ const FunctionDoc max_element_wise_doc{
     {"*args"},
     "ElementWiseAggregateOptions"};
 
+const FunctionDoc between_doc{"Check if values are in a range x <= y <= z",
+                              ("A null on either side emits a null comparison result."),
+                              {"x", "y", "z"}};
+
 }  // namespace
 
 void RegisterScalarComparison(FunctionRegistry* registry) {
@@ -767,6 +840,11 @@ void RegisterScalarComparison(FunctionRegistry* registry) {
   auto max_element_wise =
       MakeScalarMinMax<Maximum>("max_element_wise", max_element_wise_doc);
   DCHECK_OK(registry->AddFunction(std::move(max_element_wise)));
+}
+
+void RegisterScalarBetween(FunctionRegistry* registry) {
+  auto between = MakeBetweenFunction<Between>("between", &between_doc);
+  DCHECK_OK(registry->AddFunction(std::move(between)));
 }
 
 }  // namespace internal
