+++ cpp/src/arrow/compute/kernels/scalar_compare_benchmark.cc
@@ -98,5 +98,42 @@ BENCHMARK(MaxElementWiseArrayScalarInt64)->Apply(RegressionSetArgs);
 BENCHMARK(MaxElementWiseArrayArrayString)->Apply(RegressionSetArgs);
 BENCHMARK(MaxElementWiseArrayScalarString)->Apply(RegressionSetArgs);
 
+template <typename Type>
+static void BetweenScalarArrayScalar(benchmark::State& state) {
+  RegressionArgs args(state, /*size_is_bytes=*/false);
+  auto ty = TypeTraits<Type>::type_singleton();
+  auto rand = random::RandomArrayGenerator(kSeed);
+  auto array = rand.ArrayOf(ty, args.size, args.null_proportion);
+  auto scalar_left = *rand.ArrayOf(ty, 1, 0)->GetScalar(0);
+  auto scalar_right = *rand.ArrayOf(ty, 1, 0)->GetScalar(0);
+  for (auto _ : state) {
+    ABORT_NOT_OK(CallFunction("between", {array, scalar_left, scalar_right}).status());
+  }
+}
+
+template <typename Type>
+static void BetweenArrayArrayArray(benchmark::State& state) {
+  RegressionArgs args(state, /*size_is_bytes=*/false);
+  auto ty = TypeTraits<Type>::type_singleton();
+  auto rand = random::RandomArrayGenerator(kSeed);
+  auto lhs = rand.ArrayOf(ty, args.size, args.null_proportion);
+  auto mid = rand.ArrayOf(ty, args.size, args.null_proportion);
+  auto rhs = rand.ArrayOf(ty, args.size, args.null_proportion);
+  for (auto _ : state) {
+    ABORT_NOT_OK(CallFunction("between", {mid, lhs, rhs}).status());
+  }
+}
+
+static void BetweenArrayArrayArrayInt64(benchmark::State& state) {
+  BetweenArrayArrayArray<Int64Type>(state);
+}
+
+static void BetweenScalarArrayScalarInt64(benchmark::State& state) {
+  BetweenScalarArrayScalar<Int64Type>(state);
+}
+
+BENCHMARK(BetweenArrayArrayArrayInt64)->Apply(RegressionSetArgs);
+BENCHMARK(BetweenScalarArrayScalarInt64)->Apply(RegressionSetArgs);
+
 }  // namespace compute
 }  // namespace arrow
