+++ cpp/src/arrow/compute/registry.cc
@@ -278,7 +278,6 @@ static std::unique_ptr<FunctionRegistry> CreateBuiltInRegistry() {
   RegisterScalarBoolean(registry.get());
   RegisterScalarCast(registry.get());
   RegisterScalarComparison(registry.get());
-  RegisterScalarFillNull(registry.get());
   RegisterScalarIfElse(registry.get());
   RegisterScalarNested(registry.get());
   RegisterScalarRandom(registry.get());  // Nullary
@@ -288,9 +287,6 @@ static std::unique_ptr<FunctionRegistry> CreateBuiltInRegistry() {
   RegisterScalarTemporalBinary(registry.get());
   RegisterScalarTemporalUnary(registry.get());
   RegisterScalarValidity(registry.get());
-  RegisterScalarIfElse(registry.get());
-  RegisterScalarTemporalBinary(registry.get());
-  RegisterScalarTemporalUnary(registry.get());
 
   RegisterScalarOptions(registry.get());
