+++ cpp/src/arrow/compute/registry_internal.h
@@ -30,7 +30,6 @@ void RegisterScalarBetween(FunctionRegistry* registry);
 void RegisterScalarBoolean(FunctionRegistry* registry);
 void RegisterScalarCast(FunctionRegistry* registry);
 void RegisterScalarComparison(FunctionRegistry* registry);
-void RegisterScalarFillNull(FunctionRegistry* registry);
 void RegisterScalarIfElse(FunctionRegistry* registry);
 void RegisterScalarNested(FunctionRegistry* registry);
 void RegisterScalarRandom(FunctionRegistry* registry);  // Nullary
@@ -40,9 +39,6 @@ void RegisterScalarStringUtf8(FunctionRegistry* registry);
 void RegisterScalarTemporalBinary(FunctionRegistry* registry);
 void RegisterScalarTemporalUnary(FunctionRegistry* registry);
 void RegisterScalarValidity(FunctionRegistry* registry);
-void RegisterScalarIfElse(FunctionRegistry* registry);
-void RegisterScalarTemporalBinary(FunctionRegistry* registry);
-void RegisterScalarTemporalUnary(FunctionRegistry* registry);
 
 void RegisterScalarOptions(FunctionRegistry* registry);
