+++ r/R/arrowExports.R
@@ -416,10 +416,6 @@ ExecNode_output_schema <- function(node) {
   .Call(`_arrow_ExecNode_output_schema`, node)
 }
 
-ExecNode_output_schema <- function(node) {
-  .Call(`_arrow_ExecNode_output_schema`, node)
-}
-
 ExecNode_Scan <- function(plan, dataset, filter, materialized_field_names) {
   .Call(`_arrow_ExecNode_Scan`, plan, dataset, filter, materialized_field_names)
 }
